+++ cmd/cli/main.go
@@ -3,17 +3,17 @@ package main
 import (
 	"news-aggregator/pkg/handler/cli"
 	"news-aggregator/pkg/model"
-	"news-aggregator/pkg/repository"
 	"news-aggregator/pkg/service"
+	"news-aggregator/pkg/storage/inmemory"
 )
 
 func main() {
-	// Initialize repository and service
+	// Initialize storage and service
 	var articles []model.Article
-	db := repository.NewInMemory(articles)
+	db := inmemory.New(articles)
 	svc := service.New(db)
 
 	// Initialize handler and execute CLI commands
 	handler := cli.NewHandler(svc)
-	handler.Execute()
+	handler.InitCommands()
 }
