+++ pkg/model/doc.go
@@ -1,5 +1,5 @@
 // Package model contains the data models for the application.
 // It contains the structs that represent the data entities in the application.
 // It provides a common data representation for the application.
-// The model package is used by the repository and service layers to interact with the data entities.
+// The model package is used by the storage and service layers to interact with the data entities.
 package model
