+++ pkg/parser/doc.go
@@ -1,5 +1,4 @@
 // Package parser implements the parser layer of the application.
 // It contains the parsing logic for the application.
-// It interacts with the model layer to parse the data entities.
 // It provides an abstraction over the parsing operations to the service layer.
 package parser
