+++ pkg/parser/html/parser.go
@@ -1,4 +1,4 @@
-package parser
+package html
 
 import (
 	"fmt"
@@ -9,14 +9,14 @@ import (
 	"time"
 )
 
-// HtmlParser is a struct that contains the configuration for parsing HTML feeds
+// Parser is a struct that contains the configuration for parsing HTML feeds
 // and implements the ParsingAlgorithm interface.
-type HtmlParser struct {
-	config HtmlFeedConfig
+type Parser struct {
+	config FeedConfig
 }
 
-// HtmlFeedConfig is a struct that contains the configuration for parsing HTML feeds.
-type HtmlFeedConfig struct {
+// FeedConfig is a struct that contains the configuration for parsing HTML feeds.
+type FeedConfig struct {
 	ArticleSelector     string
 	TitleSelector       string
 	LinkSelector        string
@@ -28,12 +28,12 @@ type HtmlFeedConfig struct {
 }
 
 // NewHtmlParser creates a new HtmlParser with the given configuration.
-func NewHtmlParser(config HtmlFeedConfig) *HtmlParser {
-	return &HtmlParser{config: config}
+func NewHtmlParser(config FeedConfig) *Parser {
+	return &Parser{config: config}
 }
 
-// parseFile parses the given file and returns a slice of articles.
-func (h *HtmlParser) parseFile(f *os.File) ([]model.Article, error) {
+// ParseFile parses the given file and returns a slice of articles.
+func (h *Parser) ParseFile(f *os.File) ([]model.Article, error) {
 	var articles []model.Article
 	doc, err := goquery.NewDocumentFromReader(f)
 	if err != nil {
