+++ pkg/parser/rss/parser.go
@@ -1,4 +1,4 @@
-package parser
+package rss
 
 import (
 	"github.com/mmcdole/gofeed"
@@ -6,11 +6,11 @@ import (
 	"os"
 )
 
-// RssParser is a struct that implements the ParsingAlgorithm interface
-type RssParser struct{}
+// Parser is a struct that implements the ParsingAlgorithm interface
+type Parser struct{}
 
-// parseFile parses the given file and returns a slice of articles.
-func (r *RssParser) parseFile(f *os.File) ([]model.Article, error) {
+// ParseFile parses the given file and returns a slice of articles.
+func (r *Parser) ParseFile(f *os.File) ([]model.Article, error) {
 	parser := gofeed.NewParser()
 	feed, _ := parser.Parse(f)
 	articles := make([]model.Article, 0)
