+++ pkg/parser/strategy/doc.go
@@ -1,0 +1,2 @@
+// Package strategy provides ability to select parsing algorithm
+package strategy
