--- pkg/repository/doc.go
@@ -1,11 +1,0 @@
-// Package repository implements the repository layer of the application.
-// It contains the database operations for the application.
-// It interacts with the model layer to perform CRUD operations on the database.
-// It provides an abstraction over the database operations to the service layer.
-//
-// The repository layer should be used by the service layer to perform database operations.
-//
-// The repository layer contains the following packages:
-//
-// - article_inmemory: contains the in-memory database implementation for the article repository
-package repository
