+++ pkg/service/article_service.go
@@ -3,7 +3,7 @@ package service
 import (
 	"fmt"
 	"news-aggregator/pkg/model"
-	"news-aggregator/pkg/repository"
+	"news-aggregator/pkg/storage"
 	"time"
 )
 
@@ -16,10 +16,10 @@ const (
 )
 
 type ArticleService struct {
-	articleRepoInMem repository.Article
+	articleStorage storage.Article
 }
 
-// Article is an interface that defines the methods for interacting with the article repository.
+// Article is an interface that defines the methods for interacting with the article storage.
 type Article interface {
 	GetAll() ([]model.Article, error)
 	GetById(id int) (model.Article, error)
@@ -30,43 +30,43 @@ type Article interface {
 	GetByDateInRange(startDate, endDate string) ([]model.Article, error)
 }
 
-func New(articleRepo repository.Article) *ArticleService {
-	return &ArticleService{articleRepoInMem: articleRepo}
+func New(articleRepo storage.Article) *ArticleService {
+	return &ArticleService{articleStorage: articleRepo}
 }
 
 // GetAll returns all articles in the database.
 func (a *ArticleService) GetAll() ([]model.Article, error) {
-	return a.articleRepoInMem.GetAll()
+	return a.articleStorage.GetAll()
 }
 
 // GetById returns the article with the given ID.
 func (a *ArticleService) GetById(id int) (model.Article, error) {
-	return a.articleRepoInMem.GetById(id)
+	return a.articleStorage.GetById(id)
 }
 
 // Create adds a new article to the database.
 func (a *ArticleService) Create(article model.Article) (model.Article, error) {
-	return a.articleRepoInMem.Create(article)
+	return a.articleStorage.Create(article)
 }
 
 // Delete removes the article with the given ID from the database.
 func (a *ArticleService) Delete(id int) error {
-	return a.articleRepoInMem.Delete(id)
+	return a.articleStorage.Delete(id)
 }
 
 // GetBySource returns all articles from the given source.
 func (a *ArticleService) GetBySource(source string) ([]model.Article, error) {
 	switch source {
 	case "abcnews":
-		return a.articleRepoInMem.GetBySource(ABCNewsSource)
+		return a.articleStorage.GetBySource(ABCNewsSource)
 	case "bbc":
-		return a.articleRepoInMem.GetBySource(BBCNewsSource)
+		return a.articleStorage.GetBySource(BBCNewsSource)
 	case "washingtontimes":
-		return a.articleRepoInMem.GetBySource(WashingtonTimesSource)
+		return a.articleStorage.GetBySource(WashingtonTimesSource)
 	case "nbc":
-		return a.articleRepoInMem.GetBySource(NBCNewsSource)
+		return a.articleStorage.GetBySource(NBCNewsSource)
 	case "usatoday":
-		return a.articleRepoInMem.GetBySource(USATodaySource)
+		return a.articleStorage.GetBySource(USATodaySource)
 	default:
 		return nil, fmt.Errorf("source not found")
 	}
@@ -74,7 +74,7 @@ func (a *ArticleService) GetBySource(source string) ([]model.Article, error) {
 
 // GetByKeyword returns all articles that contain the given keyword.
 func (a *ArticleService) GetByKeyword(keyword string) ([]model.Article, error) {
-	return a.articleRepoInMem.GetByKeyword(keyword)
+	return a.articleStorage.GetByKeyword(keyword)
 }
 
 // GetByDateInRange returns all articles published between the given start and end dates.
@@ -100,5 +100,5 @@ func (a *ArticleService) GetByDateInRange(startDate, endDate string) ([]model.Ar
 	} else {
 		endDateObj = time.Now()
 	}
-	return a.articleRepoInMem.GetByDateInRange(startDateObj, endDateObj)
+	return a.articleStorage.GetByDateInRange(startDateObj, endDateObj)
 }
